// src/api/book-service/models/api-response-book-creation-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BookCreationResponse } from './book-creation-response';

/**
 * 
 * @export
 * @interface ApiResponseBookCreationResponse
 */
export interface ApiResponseBookCreationResponse {
    /**
     * 
     * @type {BookCreationResponse}
     * @memberof ApiResponseBookCreationResponse
     */
    'data'?: BookCreationResponse;
    /**
     * 
     * @type {string}
     * @memberof ApiResponseBookCreationResponse
     */
    'message'?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiResponseBookCreationResponse
     */
    'code'?: number;
}